requirements.txt
tensorflow
keras
numpy
opencv-python
pillow
matplotlib
reportlab
tqdm
scikit-learn
pandas
